Finish TuneUp UX Form Faktur

+++ btr.application/SalesContext/FakturAgg/UseCases/GetFakturQuery.cs
@@ -25,8 +25,8 @@ namespace btr.application.SalesContext.FakturAgg.UseCases
 
         public string CustomerId { get; set; }
         public string CustomerName { get; set; }
-        public double Plafond { get; set; }
-        public double CreditBalance { get; set; }
+        public decimal Plafond { get; set; }
+        public decimal CreditBalance { get; set; }
 
         public string WarehouseId { get; set; }
         public string WarehouseName { get; set; }
